requirements.txt
apispec==6.3.0
asgiref==3.7.2
beautifulsoup4==4.12.2
bootstrap4==0.1.0
Django==4.2.7
django-bootstrap4==23.2
django-extensions==3.2.3
django-filter==23.4
django-pagination==1.0.7
django-serializer==1.2.1
django-slugify-processor==1.5.0
djangorestframework==3.14.0
future==0.18.3
gunicorn==21.2.0
marshmallow==3.20.1
packaging==23.2
paginate==0.5.6
Pillow==10.1.0
python-slugify==8.0.1
pytz==2023.3.post1
Serializer==0.2.1
serializers==0.2.4
simplejson==3.19.2
soupsieve==2.5
sqlparse==0.4.4
text-unidecode==1.3
tzdata==2023.3
